add error handlers on endpoints

+++ api.js
@@ -18,30 +18,80 @@ api.use(bodyParser.urlencoded({ extended: false }))
 api.use(bodyParser.json())
 
 // API Routes
+// GET /api/movies - Fetch all the movies on DB
 api.get('/movies', jwt(auth), async (req, res, next) => {
-  debug('GET /api/movies')
-  const movies = await Movie.find({})
+  const { user } = req
+  debug(`GET /api/movies { username: ${user} }`)
+
+  if (!user || !user.username) {
+    return next(new Error('Not Authorized'))
+  }
+
+  let movies = []
+  try {
+    movies = await Movie.find({})
+  } catch (e) {
+    return next(e)
+  }
+
   res.send(movies)
 })
 
+// POST /api/movie - Save a new movie on DB
 api.post('/movie', jwt(auth), guard.check(['movies:write']), async (req, res, next) => {
+  const { user } = req
   const movie = req.body
-  debug(`POST /api/movie ${movie}`)
-  await Movie.create(movie)
-  res.send(movie)
+  debug(`POST /api/movie ${movie} { username: ${user} }`)
+
+  if (!user || !user.username) {
+    return next(new Error('Not Authorized'))
+  }
+
+  let savedMovie = {}
+  try {
+    await Movie.create(movie)
+    savedMovie = await Movie.find({ title: movie.title })
+  } catch (e) {
+    return next(e)
+  }
+  
+  res.send(savedMovie)
 })
 
 api.get('/movie/:id', jwt(auth), async (req, res, next) => {
+  const { user } = req
   const { id } = req.params
-  debug(`GET /api/movie/${id}`)
-  const movie = await Movie.findById(id)
+  debug(`GET /api/movie/${id} { username: ${user} }`)
+
+  if (!user || !user.username) {
+    return next(new Error('Not Authorized'))
+  }
+
+  let movie = {}
+  try {
+    movie = await Movie.findById(id)
+  } catch (e) {
+    return next(e)
+  }
+
   res.send(movie)
 })
 
 api.delete('/movie/:id', jwt(auth), async (req, res, next) => {
+  const { user } = req
   const { id } = req.params
   debug(`DELETE /api/movie/${id}`)
-  await Movie.findByIdAndRemove(id)
+
+  if (!user || !user.username) {
+    return next(new Error('Not Authorized'))
+  }
+
+  try {
+    await Movie.findByIdAndRemove(id)
+  } catch (e) {
+    return next(e)
+  }
+  
   res.send({ message: `Movie with ID: ${id} has been deleted`})
 })
